# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nzila API - Angolan AI Assistant
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ChatRequest {
  message: string;
  userId: string;
  /** @nullable */
  conversationId?: number | null;
}

export interface ChatResponse {
  response: string;
  /** dataset | ai */
  source: string;
  /** @nullable */
  matchedGiria?: string | null;
  conversationId: number;
}

export interface Giria {
  id: number;
  term: string;
  definition: string;
  example: string;
  culturalContext: string;
  category: string;
}

export interface Contribution {
  id: number;
  userId: string;
  term: string;
  definition: string;
  example: string;
  status: string;
  createdAt: string;
}

export interface CreateContributionBody {
  userId: string;
  term: string;
  definition: string;
  example: string;
}

export interface ContributionResult {
  message: string;
  status: string;
  /** @nullable */
  contributionId?: number | null;
}

export interface UserRanking {
  userId: string;
  /** Confiável | Nenhum | Moderado | Horrível */
  level: string;
  approvedContributions: number;
  totalContributions: number;
  isBlocked: boolean;
}

export interface ApiError {
  error: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}