# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nzila API - Angolan AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Send a chat message to Nzila
 */
export const SendChatMessageBody = zod.object({
  message: zod.string(),
  userId: zod.string(),
  conversationId: zod.number().nullish(),
});

export const SendChatMessageResponse = zod.object({
  response: zod.string(),
  source: zod.string().describe("dataset | ai"),
  matchedGiria: zod.string().nullish(),
  conversationId: zod.number(),
});

/**
 * @summary List all gírias (Angolan slang) in the dataset
 */
export const ListGiriasResponseItem = zod.object({
  id: zod.number(),
  term: zod.string(),
  definition: zod.string(),
  example: zod.string(),
  culturalContext: zod.string(),
  category: zod.string(),
});
export const ListGiriasResponse = zod.array(ListGiriasResponseItem);

/**
 * @summary List pending contributions
 */
export const ListContributionsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  term: zod.string(),
  definition: zod.string(),
  example: zod.string(),
  status: zod.string(),
  createdAt: zod.date(),
});
export const ListContributionsResponse = zod.array(
  ListContributionsResponseItem,
);

/**
 * @summary Submit a new gíria or expression contribution
 */
export const CreateContributionBody = zod.object({
  userId: zod.string(),
  term: zod.string(),
  definition: zod.string(),
  example: zod.string(),
});

/**
 * @summary Get user ranking and stats
 */
export const GetUserRankingParams = zod.object({
  userId: zod.coerce.string(),
});

export const GetUserRankingResponse = zod.object({
  userId: zod.string(),
  level: zod.string().describe("Confiável | Nenhum | Moderado | Horrível"),
  approvedContributions: zod.number(),
  totalContributions: zod.number(),
  isBlocked: zod.boolean(),
});

/**
 * @summary List all conversations
 */
export const ListOpenaiConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.date(),
});
export const ListOpenaiConversationsResponse = zod.array(
  ListOpenaiConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateOpenaiConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get conversation with messages
 */
export const GetOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetOpenaiConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.date(),
    }),
  ),
});

/**
 * @summary Delete a conversation
 */
export const DeleteOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List messages in a conversation
 */
export const ListOpenaiMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListOpenaiMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.date(),
});
export const ListOpenaiMessagesResponse = zod.array(
  ListOpenaiMessagesResponseItem,
);

/**
 * @summary Send a text message and receive a streaming text response
 */
export const SendOpenaiMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendOpenaiMessageBody = zod.object({
  content: zod.string(),
});