# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nzila API - Angolan AI Assistant
 * OpenAPI spec version: 0.1.0
 */

export interface ChatResponse {
  response: string;
  /** dataset | ai */
  source: string;
  /** @nullable */
  matchedGiria?: string | null;
  conversationId: number;
}