# lib/api-zod/src/generated/types/contribution.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nzila API - Angolan AI Assistant
 * OpenAPI spec version: 0.1.0
 */

export interface Contribution {
  id: number;
  userId: string;
  term: string;
  definition: string;
  example: string;
  status: string;
  createdAt: Date;
}