# lib/api-zod/src/generated/types/contributionResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nzila API - Angolan AI Assistant
 * OpenAPI spec version: 0.1.0
 */

export interface ContributionResult {
  message: string;
  status: string;
  /** @nullable */
  contributionId?: number | null;
}