# lib/api-zod/src/generated/types/giria.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nzila API - Angolan AI Assistant
 * OpenAPI spec version: 0.1.0
 */

export interface Giria {
  id: number;
  term: string;
  definition: string;
  example: string;
  culturalContext: string;
  category: string;
}