# lib/api-zod/src/generated/types/userRanking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nzila API - Angolan AI Assistant
 * OpenAPI spec version: 0.1.0
 */

export interface UserRanking {
  userId: string;
  /** Confiável | Nenhum | Moderado | Horrível */
  level: string;
  approvedContributions: number;
  totalContributions: number;
  isBlocked: boolean;
}